Update blender plugin info for PRPEE
I think it is necessary to update tracker_url and wiki_url to be able to access the repository from the blender plugin manager. I also updated the name, description and added your Github name, I apologize if you find this wrong. Thank you very much for continuing with the project, it is very useful.